Added a new AlphaTrie class that maps alphabetical strings (case-insensitive) to a type T for that instance. Lookup is at most O(k), where k is the longest length string in the Trie. Insertion is also at most O(k).
AlphaTrie is used to find a certain string or find the best completion for a string (such as 'loo'->'look')

AlphaTrie still needs a Remove function to be implemented before it can be used for the next logical step (explained below.)

Changed the Player class to use an AlphaTrie to keep track of the commands available to that player. Even though it only takes microseconds in order to construct their personal trie, as the number of commands increases, it'd probably be wise to keep a static instance of a base starting command list and create a copy constructor for AlphaTrie, then when a player logs on, copying that AlphaTrie over to their personal instance.

Changed the commands to map command names to delegates, and implemented the commands as such. Now process input checks the trie for an appropriate command and if one is found that delegate is called.

The next step, after Remove is added (which won't take much) is to set it so that exit names in a room are alphabetical only, and when you enter a room, the room automatically adds those exit names to your commands list. Then, if there is an exit named 'north' in the room, for example, you'll automatically have a 'north' command to use that maps to it. Go <x> will still be used for Go'ing to an exit based on the description. When you exit a room, it'll appropriately remove the old exit commands.

Additionally, this means that each room or even each entity in a room can have their own commands they grant the players in the room (for example a room that has a level can have a 'pull' command coded specifically for that room.)

Finally, it can be used for qualifying the player for special commands like Guild commands or Admin commands, where only if they have those flags will those commands be added to their repertoire.

AlphaTrie doesn't do any explicit Disposal, but I don't think that the GC would have a difficult time keeping up with the references if an AlphaTrie goes out of scope.

I'd also like to start a MUDServer test suite, especially for AlphaTrie testing. Might do that this weekend.

Some other ideas for AlphaTrie include using it for the connected players list, so that if you do 'tell sekk hi' it will map sekk-> Sekkusu if that player is connected.