Changed AlphaTrie.FindByStartKey to use an enumerator to hand out the values that it finds. AlphaTrie.FindByExactKey then uses this enumerator to look for the exact key in the returned list.
Commented out Remove temporarily until I figure out exactly what to do with it. Ideally I'd like to include some sort of recompression in the Trie after a node is deleted, but I don't know how easy that will be.

Added a new test for AlphaTrie that checks FindByStartKey (and incidentally Insert.)